+++ refactor-this/refactor-this/Models/Transaction.cs
@@ -8,11 +8,11 @@ namespace refactor_this.Models
 {
     public class Transaction
     {
-        public float Amount { get; set; }
+        public Decimal Amount { get; set; }
 
         public DateTime Date { get; set; }
 
-        public Transaction(float amount, DateTime date)
+        public Transaction(Decimal amount, DateTime date)
         {
             Amount = amount;
             Date = date;
